addrdataset() now returns the existing rdataset in addedrdataset if the result is DNS_R_UNCHANGED.
Adjust find_deepest_zonecut() to deal with the new treatment of level_matches
by the RBT code.  These changes, combined with the RBT changes, fix a bug
where the wrong delegation was returned if the partial match was a terminal
node in the RBT.

Deal with wildcards of the form *.<zone_origin>.  This stops us from
treating the zone top as a delegation point if wildcards are present.

delegating_type() was erroneously returning true for type NS records in a
cache database.