duplicate log message filtering functionality via isc_log_write1() and isc_log_vwrite1(), configurable over X seconds using isc_log_setduplicateinterval().
isc_log_vwrite renamed to static isc_log_doit(), isc_log_vwrite() created
as wrapper for isc_log_doit().

don't do a second test on ISC_LOG_PRINTLEVEL when logging the message to add
a colon-space, since the colon-space is already in level_string.

print "no_module: " if a module has not been specified (but PRINTMODULE was),
to keep tokenization nice and clean for future log-file parsing programs
without looking quite like a bug, which is what happens with:
  Oct 25 15:51:07 dns_general: : notice: This should be in file 1/1
instead of
  Oct 25 15:51:07 dns_general: no_module: notice: This should be in file 1/1