been having trouble with the FreeeBSD system - wait, waitpid and family act totally broken, but only for programs run from cron. the waitpid loop in t_api.c times out eventually, and the testcases are killed and reaped, but the net effect of the problem is to cause each testcase to take 60 seconds, which is the default timeout. As a result, the tests take forever to run and often don't finish before the report generator is run, causing the report generator to incorrectly reports an overall test status based on an incomplete journal ...
so ... I've added some more simplistic journal correctness checking
for now so that the FreeBSD test status is correctly reported - if
a journal doesn't end with the correct end stanza, a test result of
inspect(incomplete journal) is reported.

obviously, at some point in the future, a test case count needs added
so that the report generator can do more complete correctness checking
of the journal.