Cancelation of TCP queries while they were still connecting was broken, and would cause seg faults.
Do not update the RTT if a query is being canceled due to internal failures.

Some servers generate badly formatted responses when they get an EDNS
query.  We were marking these servers as bad, but a more practical
solution is to retry without EDNS.  If a message fails to parse due to
DNS_R_FORMERR or DNS_R_UNEXPECTEDEND, and we were using EDNS, we now
retry the query without EDNS.

Add a "default" case to the message parsing error switch.  This prevents bad
things from happening if message parsing fails in a nontypical way.