Print lock tracing to stderr.
Decrement granted read lock when releasing a read lock if there are no
pending write locks.  This still allows fair queuing between read and
write locks but does not block on aquiring a new read relock if we are
below quota of active locks and there are no pending writes.