The 'forwarders' and 'forward' options are now hooked up.
If the server had been bound to a particular query source and was then
reloaded, the reloading would fail due to "address in use" if the query
source had not been changed.  The server now remembers if a custom dispatcher
is in use, and reuses it if it is compatible with the new configuration.