isc_logchannel, isc_logchannellist, and isc_logmessage now use the isc list macros.
ISC_LOG_DEBUGONLY flag for a channel added, to mark a channel that should
only be used if the debugging level is non-zero.

Use ISC_LOG_DEBUGONLY for default_debug.

REQUIRE the context is valid for isc_logconfig_get, and ENSURE that
the configuration returned is non-NULL.

Other miscellaneous cleanups.