According to the latest KAME release, recently added to NetBSD, it isn't permitted to directly include netinet6/in6.h anymore.  The RFCs and drafts all say to include netinet/in.h and that will define the ipv6 goop if the host supports it.
Rather than break (broken) OSs, I added a check for NetBSD in configure
that will force the definitions of

	ISC_PLATFORM_HAVENETINET6IN6H
and	LWRES_PLATFORM_HAVENETINET6IN6H

to be #undef.

This isn't really as clean as I'd like.  I'd rather be hard and say an OS
that requires direct include of netinet6/in6.h is broken, and remove it
fully from our code.