Silence IRIX warnings: "adb.c", line 2042: remark(1552): variable "result" was set but never used         isc_result_t result;                      ^
"adb.c", line 3082: warning(1185): enumerated type mixed with another type
                               NAME_GLUEOK(adbname), NAME_HINTOK(adbname),
                                                     ^

"adb.c", line 3187: warning(1185): enumerated type mixed with another type
                               now, NAME_GLUEOK(adbname), NAME_HINTOK(adbname),

Note that the first is in timer_cleanup, a function returning void in a context
that can't do anything with something other than void (as an isc_taskaction_t).
It calls isc_timer_reset, which could conceivably return ISC_R_NOMEMORY or
ISC_R_UNEXPECTED.  For now the call has been cast to void, but someone
who knows the code better should have a look and see whether something more
intelligent could be done with an error (and, if not, remove my XXX and
make sure the comment is clear as to why ignoring the return is the
best that can be done).