285.	[bug]		A change made to the dst API for beta4 inadvertently 			broke OMAPI's creation of a dst key from an incoming 			message, causing an assertion to be triggered.  Fixed.
Also fixed was a problem where a failed call to dns_name_fromtext would
have meant the mutex was not unlocked.  This was fixed in the mainline too,
I just forgot to mention it.

A CHANGES entry was not committed because it isn't clear to me how we are
handling things changed on the mainline but not under other release tags.
The bind9_v9_0 CHANGES file is missing #283 and #284 from the mainline.