DO NOT PULL ME UP TO THE RELEASE BRANCH.
This is a fix for the race condition in dig.  It seems to work well, except
there's a memory leak I can't quite track down.  Michael and I will look at
it on the plane ride tommorow, and commit a change via modem once we arrive
in PA.  We'll also be doing general code cleanup.