I'm apparently good at confusing CVS.  This diff has no changes.
The previous one, on the other hand (with an empty log message), is a fairly
large rewrite of nsupdate.  It's much more compatible with the bind 8 version
(the commands in the man page actually work now).  There's a lot less code
duplication, and some other general simplification.  This is still a work in
progress.