Bug #141
dig -b now tries its best to use the address you tell it to bind to.  The
problem was telling it to bind to an address in one family, then connect to
a server in a different one.  Now, it only tries servers in the same family
you bind to (if you specify -b) and gives a more meaningful error message
if it can't.