519.	[bug]		dns_name_split() would improperly split some bitstring 			labels, zeroing a few of the least signficant bits in 			the prefix part.  When such an improperly created 			prefix was returned to the RBT database, the bogus 			label was dutifully stored, corrupting the tree. 			[RT #369]
Also made dns_name_split() REQUIRE that suffixlabels always be greater than 0,
even when splitting a bitstring label (it already required this when not
splitting a bitstring label).  This is consistent with the way dns_name_split()
was called to split a name that consisted of a single label, a bitstring;
the appropriate suffixlabels value is 1 in such cases.

Also a fixed minor style error, and a confusing comment.