534.	[func]		Ancestors have been removed from RBT chains.  Ancestor 			information can be discerned via node parent pointers.
 533.	[func]		Incorporated name hashing into the RBT database to
			improve search speed.

There is still evidence of a bug with regard to bitstring labels.  It shows
up in bin/test/rbt/t_rbt -x -t 4 when the assertion at lib/dns/rbt.c:1631
is uncommented -- essentially a bitstring node's location in the hashtable
is not getting properly updated at some point.  This shouldn't affect
searching, because a bitstring label as the parent of a new level will
generally cause the standard old binary search to be done.  I will be looking
at this more closely in the very near future.