Make greatest_version() return an isc_result_t, which it should have been doing all along since there was the possibility isc_dir_open() would fail and overloading the return value of "0" was bad practice.
If isc_dir_open() fails in greatest_version(), make sure the path separator
is restored to the destination pathname if it was NULified before returning
the error.

The underlying problem of isc_log_open() not having the bad result of
isc_dir_open() reported anywhere will be slightly mediate by a change
to bin/named/logconf.c to test the directory.  lib/isc/log.c can't really
handle it because it has no idea to where such an error should be reported,
and it only shows up during isc_log_open() as part of the isc_log_write()
family of calls, which are all void.