The most notable change is that incremental_cleaning_action() will now wait at least six seconds between each pass of the tree.  This keeps the CPU from being bogged down in a persistent overlimit condition, and because it uses timers also lets the cache cleaner know right away when the server is being shut down, so it can stop trying to clean.
Also fixed an unlocked access to cache->live_tasks in dns_cache_detach().

Also fixed a problem with repeated calls to dns_cache_setcachesize() setting
the database as not overmem, even though it might have still been in
the process of cleaning from a previous limit.  Just set the new
isc_mem_setwater() limits instead and let the mem.c call water() if cleaning
needs to be either started or stopped because of the new limits.

Minor code cleanups.