2600.  [doc]           ARM: miscellaneous reformatting for different                        page widths. [RT #19574]
Split comments to multiple lines or move comments to own lines for
example.

Some fixed width examples are too wide for my printed page, so fixed

Splits many programlisting's to multiple lines.
Some move comments to above and some merge some comments.

Note that this patch covers configuration examples and also the "Grammar".

Also while here fix docbook formatting for a Not Implemented
check-names (but still that is too wide so I need to change).  And
also the patch changes a couple tabs to spaces. And changed case of
example comments to be consistent with previous lines.

There shouldn't be any content changes here.

This doesn't complete this PR.
Even this patch, the official ARM PDF has other text that disappears
into the right margin. So still working on this.