Update grammar for sig-validity-interval in ARM to include optional second argument.
I had noticed a few days ago that the ARM grammar for sig-validity-interval
was missing the optional second number. I saw an email again about
this today. This is for part of my RT #19931.

Not adding a CHANGES entry for this (because feature itself was
already documented.)

This is for v9_6. (Already did HEAD.)