Remove canary code from hash destroy function (#38602)
This triggers a Valgrind out-of-bounds read report. It was introduced by
commit 5d7849ad7ffc6d08870dbfbc8d6bfffd90007488.

No CHANGES entry necessary as it doesn't have any user-visible or
behavioral change. It removes an out-of-bounds read issue that went
undetected when allocated through isc_mem as the memory was present.
The memory read was compared to itself, so it has no behavioral change.