remove the only section in the history appendix to workaround dblatex issue
dblatex generated LaTeX that failed with the "id" for link reference
in the sect1 when itr had no <title> to reference.
(A workaround is to set a <title>.)
But since this appendix only had one section
and looked off to have B1 with no title and no B2, just remove the
sect1 tags.
I added the id to the first <para> tag just in case someone uses
it to link.
Note I didn't reformat the content there.

I didn't get this reviewed as was trivial change.