move nosit-udp-size doc
This is for ticket #39570.
I didn't have this reviewed for v9_10, but was reviewed for the
master but had different option names.

Moved some docs from server statement to options.
While there add a missing related option from grammar.
Also improved docs a little for nosit-udp-size.

Note I didn't add a changelog entry. The documentation was already
there and only modified a little.