Extract the do-while loop in update_sigs() into a separate function
The do-while loop in update_sigs() is hard to follow due to heavy
nesting and the 'tuple' variable also being used in the outer for loop.
Add a comment to explain the purpose of the do-while loop.  Extract it
into a separate function to decrease indentation and prevent using
'tuple' in two different loops.