Improve the way cache contents are searched for "ns.flushtest.example"
Expecting "ns.flushtest.example" to always be the first name in the ADB
dump is fragile, because in a certain corner case "a.root-servers.nil"
will be the first name instead.

As the purpose of the relevant check is to ensure "ns.flushtest.example"
is removed from ADB by "rndc flushtree flushtest.example", search the
entire list of names present in ADB instead of just the first entry when
looking for "ns.flushtest.example".