Alter distclean rule to prevent recursing into the same directory twice
Also, restore the "system" directory in bin/tests/Makefile.in to both
SUBDIRS and TESTDIRS variables. It is needed in SUBDIRS for the build,
and in TESTDIRS so "amke test" will work. Changing the distclean rule
to check for the existence of a Makefile in each directory before
recursing prevents it from attempting to run in a directory where it
has already run.