copy_name(): only allocate a buffer for non-zero length RDATA
The RDATA part of a resource record can be empty, which means
copy_name() may use a { .base = NULL, .length = 0 } region with
isc_buffer_putmem().  While harmless, it triggers an ubsan warning:

    client.c:2625:4: runtime error: null pointer passed as argument 2, which is declared to never be null

This happens because isc_buffer_putmem() is a macro which uses
memmove(), which in turn expects its second argument (the source
pointer, i.e. the region's base address) to never be NULL.

Only allocate and subsequently process a buffer if RDATA has non-zero
length to prevent the warning from being triggered.