Implement a dns_zone_log*() variant taking a va_list argument
This function will be useful for the reworked dns_zoneverify_dnssec(),
which will use a helper function for logging that will either print to
stderr or pass its variable arguments to dns_zone_log*().

Rework dns_zone_log() and dns_zone_logc() so that they make use of the
new function, which decreases code duplication.